Move original ChadoConnection functional tests over to the new kernel test.
